Generate Circle CI config file version 2 Prompting user to override if old config exists

+++ generators/app/index.js
@@ -7,33 +7,35 @@ module.exports = class extends Generator {
     // Have Yeoman greet the user.
     this.log(yosay('Welcome to the JHipster CircleCI 2 configuration file generator!'));
 
-    const prompts = [
-      {
-        type: 'confirm',
-        name: 'someAnswer',
-        message: 'Would you like to enable this option?',
-        default: true
-      }
-    ];
+    if (this.fs.exists(this.destinationPath('.circleci/config.yml'))) {
+      const prompts = [
+        {
+          type: 'confirm',
+          name: 'override',
+          message: 'Would you like to override existing Circle CI configuration file?',
+          default: true
+        }
+      ];
 
-    return this.prompt(prompts).then(props => {
-      // To access props later use this.props.someAnswer;
-      this.props = props;
-    });
+      return this.prompt(prompts).then(props => {
+        // To access props later use this.props.someAnswer;
+        this.props = props;
+      });
+    } else {
+      this.props = {
+        override: true
+      };
+    }
   }
 
   writing() {
-    this.fs.copy(
-      this.templatePath('dummyfile.txt'),
-      this.destinationPath('dummyfile.txt')
-    );
+    if (this.props.override) {
+      this.fs.copy(
+        this.templatePath('circleci.yml'),
+        this.destinationPath('.circleci/config.yml')
+      );
+      this.log('Circle CI configuration file has been created successfully.');
+    };
   }
 
-  // A install() {
-  //   this.installDependencies({
-  //     npm: false,
-  //     bower: false,
-  //     yarn: false
-  //   });
-  // }
 };
